execution  stage editor updated
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C31460

+++ TOneV2/Code/TOneV2/TOne.WhS.SupplierPriceList.BP.Activities/GenerateRatesPreview.cs
@@ -38,7 +38,7 @@ namespace TOne.WhS.SupplierPriceList.BP.Activities
                         ZoneName = item.ZoneName,
                         ChangeType = item.ChangeType,
                         RecentRate = item.ProcessInfo.RecentRate,
-                        NewRate = item.NormalRate,
+                        NewRate =(decimal) item.NormalRate,
                         BED = item.BED,
                         EED = item.EED
                     });
